+++ demosplan/DemosPlanCoreBundle/Controller/Procedure/DemosPlanProcedureController.php
@@ -2586,6 +2586,7 @@ class DemosPlanProcedureController extends BaseController
                 'form'                         => $form,
                 'boilerplateCategories'        => $boilerplateCategories,
                 'boilerplateGroupsOfProcedure' => $boilerplateGroups,
+                'boilerplateUsages'            => $this->procedureService->getBoilerplateUsagesForDisplay($boilerplateId),
                 'selectedGroup'                => '',
                 'title'                        => 'procedure.boilerplate.edit',
                 'procedure'                    => $procedure,