+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+### Added
+- Boilerplate edit page lists the segments in which a boilerplate is used as deep links
 
 ### Added
 - Tags can now be linked to a default assignee, who is preselected when the tag is added to a segment during statement splitting